+++ modules/indicia_svc_import/controllers/services/import.php
@@ -418,12 +418,24 @@ class Import_Controller extends Service_Base_Controller {
             }
             $index++;
           }
+          // The genus, specific name and qualifier are all merge fields.
+          // However the qualifier is not mandatory, so if a qualifier is not specified, we effectively tell the system it has been specified
+          // so that the system doesn't ask for it. Ideally this code should be generalised going forward.
+          if ((in_array('taxon:taxon:genus',$metadata['mappings']) || in_array('taxon:taxon:specific',$metadata['mappings'])) &&  !in_array('taxon:taxon:qualifier',$metadata['mappings'])) {
+            $columns['taxon:taxon:qualifier'] = TRUE;
+          }
           if (count($defn['columns']) === count(array_keys($columns))) {
             $specialFieldProcessing[$column] = TRUE;
           }
         }
       }
       $specialMergeProcessing = array();
+      if (isset($metadata['importMergeFields']) && is_string($metadata['importMergeFields'])) {
+        $metadata['importMergeFields'] = json_decode($metadata['importMergeFields'], TRUE);
+      }
+      if (isset($metadata['synonymProcessing']) && is_string($metadata['synonymProcessing'])) {
+        $metadata['synonymProcessing'] = json_decode($metadata['synonymProcessing'], TRUE);
+      }
       if (isset($metadata['importMergeFields'])) {
         // Only do the special merge processing if all the required fields are
         // there, and if there are no required then if one of the optional ones
@@ -480,6 +492,12 @@ class Import_Controller extends Service_Base_Controller {
             }
             $index++;
           }
+        }
+        // The genus, specific name and qualifier are all merge fields.
+        // However the qualifier is not mandatory, so if a qualifier is not specified, we effectively tell the system it has been specified
+        // so that the system doesn't ask for it. Ideally this code should be generalised going forward.
+        if ((array_key_exists('taxon:taxon:genus',$saveArray) || array_key_exists('taxon:taxon:specific',$saveArray)) &&  !array_key_exists('taxon:taxon:qualifier',$saveArray)) {
+          $saveArray['taxon:taxon:qualifier'] = '';
         }
         foreach (array_keys($specialFieldProcessing) as $col) {
           if (!isset($saveArray[$col]) || $saveArray[$col] == '') {
@@ -830,7 +848,7 @@ class Import_Controller extends Service_Base_Controller {
       $this->auto_render = FALSE;
       if (!empty($allowCommitToDB)&&$allowCommitToDB==true) {
         $cache->set(basename($csvTempFile) . 'previousSupermodel', $this->previousCsvSupermodel);
-      }       
+      }
       if (class_exists('request_logging')) {
         request_logging::log('i', 'import', NULL, 'upload',
           empty($saveArray['website_id']) ? NULL : $saveArray['website_id'],
@@ -1104,7 +1122,7 @@ class Import_Controller extends Service_Base_Controller {
               // Taxon info may not be provided if looking up existing record.
               // In which case, skip the lookup.
               !empty($saveArray['taxon:language_id']) &&
-              (!empty($saveArray['taxon:taxon']) || !empty($saveArray['taxon:external_key']))) {
+              (!empty($saveArray['taxon:taxon']) || !empty($saveArray['taxon:external_key']  || !empty($saveArray['taxon:search_code'])))) {
             // Same for taxa_taxon_lists, and their taxon supermodel: have to
             // look up using complex query to get the link between the
             // taxon_list and the taxon.
@@ -1134,6 +1152,10 @@ class Import_Controller extends Service_Base_Controller {
             }
             elseif (in_array('taxon:external_key', $fields) && isset($saveArray['taxon:external_key'])) {
               $query .= "AND t.external_key ='" . $saveArray['taxon:external_key'] . "' ";
+              $existing = $db->query($query)->result_array(FALSE);
+            }
+            elseif (in_array('taxon:search_code', $fields) && isset($saveArray['taxon:search_code'])) {
+              $query .= "AND t.search_code ='" . $saveArray['taxon:search_code'] . "' ";
               $existing = $db->query($query)->result_array(FALSE);
             }
             else {