+++ modules/rest_api/controllers/services/rest.php
@@ -72,6 +72,15 @@ class RestApiAbort extends Exception {}
 class Rest_Controller extends Controller {
 
   /**
+   * Defines which ES CSV column download template to use.
+   *
+   * Only supports "default" or empty string currently.
+   *
+   * @var string
+   */
+  private $esCsvTemplate = 'default';
+
+  /**
    * Set sensible defaults for the authentication methods available.
    *
    * @var array
@@ -719,10 +728,8 @@ class Rest_Controller extends Controller {
    * Template for ES CSV output.
    *
    * @var array
-   *
-   * @todo Make this configurable.
    */
-  private $esCsvTemplate = [
+  private $defaultEsCsvTemplate = [
     'Record ID' => 'id',
     'RecordKey' => '_id',
     'Sample ID' => 'event.event_id',
@@ -790,12 +797,28 @@ class Rest_Controller extends Controller {
     // Either unscrolled, or the first call to a scroll. So post the query.
     $postData = file_get_contents('php://input');
     $postObj = empty($postData) ? [] : json_decode($postData);
+    // Params for configuring an ES CSV download template get extracted and not
+    // sent to ES.
+    if (isset($postObj->columnsTemplate)) {
+      $this->esCsvTemplate = $postObj->columnsTemplate;
+      unset($postObj->columnsTemplate);
+    }
+    if (isset($postObj->addColumns)) {
+      $this->esCsvTemplateAddColumns = (array) $postObj->addColumns;
+      unset($postObj->addColumns);
+    }
+    if (isset($postObj->removeColumns)) {
+      $this->esCsvTemplateRemoveColumns = (array) $postObj->removeColumns;
+      unset($postObj->removeColumns);
+    }
 
     if ($scrollMode !== 'off') {
       $postObj->size = MAX_ES_SCROLL_SIZE;
     }
     if ($format === 'csv') {
-      foreach ($this->esCsvTemplate as $field) {
+      $csvTemplate = $this->getEsCsvTemplate();
+      $fields = [];
+      foreach ($csvTemplate as $field) {
         if (strpos($field, '_') === 0) {
           // Fields starting _ are not inside _source object.
           continue;
@@ -956,6 +979,7 @@ class Rest_Controller extends Controller {
    */
   private function getEsOutputHeader($format) {
     if ($format === 'csv') {
+      $csvTemplate = $this->getEsCsvTemplate();
       $row = array_map(function ($cell) {
         // Cells containing a quote, a comma or a new line will need to be
         // contained in double quotes.
@@ -964,10 +988,32 @@ class Rest_Controller extends Controller {
           return '"' . preg_replace('/"/', '""', $cell) . '"';
         }
         return $cell;
-      }, array_keys($this->esCsvTemplate));
+      }, array_keys($csvTemplate));
       return implode(',', $row) . "\n";
     }
     return '';
+  }
+
+  /**
+   * Determines the columns template for an ES download.
+   *
+   * @return array
+   *   List of column definitions to download.
+   */
+  private function getEsCsvTemplate() {
+    // Start with the default columns set, or an empty array.
+    $csvTemplate = $this->esCsvTemplate === 'default' ? $this->defaultEsCsvTemplate : [];
+    // Append extra columns.
+    if (!empty($this->esCsvTemplateAddColumns)) {
+      $csvTemplate = array_merge($csvTemplate, $this->esCsvTemplateAddColumns);
+    }
+    // Remove any that need to be removed.
+    if (!empty($this->esCsvTemplateRemoveColumns)) {
+      foreach ($this->esCsvTemplateRemoveColumns as $col) {
+        unset($csvTemplate[$col]);
+      }
+    }
+    return $csvTemplate;
   }
 
   /**
@@ -1137,9 +1183,10 @@ class Rest_Controller extends Controller {
     if (empty($data['hits']['hits'])) {
       return;
     }
+    $esCsvTemplate = $this->getEsCsvTemplate();
     foreach ($data['hits']['hits'] as $doc) {
       $row = [];
-      foreach ($this->esCsvTemplate as $source) {
+      foreach ($esCsvTemplate as $source) {
         $this->copyIntoCsvRow($doc, $source, $row);
       }
       $row = array_map(function ($cell) {
@@ -2541,6 +2588,7 @@ class Rest_Controller extends Controller {
       }
     }
     if (!$this->authenticated) {
+      // Either the authentication wrong, or using HTTP instead of HTTPS.
       $this->apiResponse->fail('Unauthorized', 401, 'Unable to authorise');
     }
   }